refactor: Change setNumberFromStr behavior
Before this commit, this function returned total number of
characters that starts from 'str' pointer. For example, if 'str'
will be '12.25.65abc' it will return 8 as number len (12.25.65), but the real
value is (12.25).

To resolve the problem we need to know separately the length of the
integer part and then, if it exists, the fractional part of the given
number.